art-service/requirements.txt
aio-pika==9.4.3
aiormq==6.8.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
black==24.8.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
google-api-core==2.19.1
google-auth==2.34.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.64.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
loguru==0.7.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
pika-stubs==0.1.3
pillow==10.4.0
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rsa==4.9
shortuuid==1.0.13
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
uvicorn==0.30.6
yarl==1.9.4
